Refactored and renamed SignificantSpaceParsers
SignificantSpaceParsers:
* was BaseParsers;
* new general method stringParser:
* logit method now adds failure parser in LogOff case;
RationalParser: parse removed and replaced by calls to stringParser;
MillParser: uses stringParser instead of parse;
NumberParser: uses stringParser instead of parse;
ShuntingYardParser: cosmetics, scaladoc
