+++ src/main/scala/com/phasmidsoftware/number/parse/ShuntingYardParser.scala
@@ -22,17 +22,27 @@ class ShuntingYardParser extends MillParser {
     * @param w the String to parse.
     * @return a Mill, wrapped in Try.
     */
-  def parseInfix(w: String): Try[Mill] = {
-    val triedShuntingYard = parseAll(shuntingYard, w) match {
-      case Success(t, _) => scala.util.Success(t)
-      case Failure(m, _) => scala.util.Failure(RationalParserException(m))
-      case Error(m, _) => scala.util.Failure(RationalParserException(m))
-    }
-    triedShuntingYard.flatMap(_.toMill)
-  }
+  def parseInfix(w: String): Try[Mill] = stringParser(shuntingYard, w).flatMap(_.toMill)
 
+  /**
+    * A ShuntingYard consisting of two stacks: the values stack and the operators stack.
+    *
+    * @param values    a list of values and operators. The former are added directly to values.
+    *                  However, operators are only added indirectly, after a closing parenthesis
+    *                  (or the end of the string) is reached.
+    * @param operators a list of operators which is temporarily placed here until switch is called.
+    */
   case class ShuntingYard(values: Seq[Item], operators: Seq[Item]) {
 
+    /**
+      * Method to add a token to this ShuntingYard.
+      *
+      * @param token the token to add.
+      * @return a new ShuntingYard which is the same as this but with token added.
+      *         If token is an operator, it is added to the operators;
+      *         If token is a number, it is added to the values;
+      *         If token is a open or close parenthesis, it is handled specially.
+      */
     def :+(token: InfixToken): ShuntingYard = token match {
       case InfixToken(Some(t), _) => t match {
         case Left(operator) => this :+ operator
@@ -43,6 +53,12 @@ class ShuntingYardParser extends MillParser {
         else switch
     }
 
+    /**
+      * Convert this ShuntingYard into a Mill.
+      * Any remaining operators must first be switched to the values stack.
+      *
+      * @return a Try[Mill].
+      */
     def toMill: Try[Mill] = switch match {
       case ShuntingYard(values, Nil) => Try(Mill(values: _*))
       case _ => scala.util.Failure(MillException("toMill: logic error"))
@@ -61,22 +77,57 @@ class ShuntingYardParser extends MillParser {
   }
 
   object ShuntingYard {
+    /**
+      * Create a new, empty, ShuntingYard.
+      *
+      * @return a new, empty, ShuntingYard.
+      */
     def apply(): ShuntingYard = ShuntingYard(Nil, Nil)
 
-    def apply(tokens: List[InfixToken]): ShuntingYard = tokens.foldLeft(ShuntingYard())(_ :+ _).switch
+    /**
+      * Create a new ShuntingYard with the tokens added.
+      *
+      * @param tokens a list of InfixToken.
+      * @return
+      */
+    def apply(tokens: Seq[InfixToken]): ShuntingYard = tokens.foldLeft(ShuntingYard())(_ :+ _).switch
   }
 
+  /**
+    * An infix token which represents either a Token or an open/close parenthesis.
+    *
+    * @param to    an optional token.
+    * @param paren the to is None, then paren is interpreted as Open (for true) and Close (for false).
+    */
   case class InfixToken(to: Option[Token], paren: Boolean)
 
+  /**
+    * Parser for a shunting yard.
+    * It matches on a list of infixTokens separated by white space.
+    *
+    * @return a Parser[ShuntingYard].
+    */
   def shuntingYard: Parser[ShuntingYard] = repSepSp(infixToken) :| "shuntingYard" ^^ (tokens => ShuntingYard(tokens))
 
-  def infixToken: Parser[InfixToken] = (maybeNumber ?| operator) ^^ {
+  /**
+    * Parser for an InfixToken.
+    * It matches on a Number, or an operator--including ( and ).
+    *
+    * @return a Parser[InfixToken].
+    */
+  def infixToken: Parser[InfixToken] = (maybeNumber ?| operator) :| "infixToken" ^^ {
     case Left(x) => InfixToken(Some(Right(x)), paren = false)
     case Right("(") => InfixToken(None, paren = true)
     case Right(")") => InfixToken(None, paren = false)
     case Right(w) => InfixToken(Some(Left(w)), paren = false)
   }
 
+  /**
+    * Parser for a String.
+    * It matches on one of the operators defined in MillParser or an open or close parenthesis.
+    *
+    * @return Parser[String].
+    */
   def operator: Parser[String] = (dyadicOperator | monadicOperator | anadicOperator | neutralOperator2 | openParenthesis | closeParenthesis) :| "operator"
 
   val openParenthesis: String = "("
